+++ rosebud/CHANGELOG.md
@@ -1,5 +1,15 @@
 # Change log
 
+## 0.2.0.0
+
+* Improve performance and ergonomics of `eitherTreeFromLabels`: root label is
+  now passed in explicitly
+* Add `eitherNEForestFromPartitionedLabels` and
+  `unsafeNEForestFromPartitionedLabels`
+* Revamp error types:
+  * Replace `NEForestFromLabelsError` and `TreeFromLabelsError` with
+    `FromPartitionedLabelsError` and `FromLabelsError`
+
 ## 0.1.0.0
 
 * Initial release